Update main.cpp
A few of the operations were commented out so I removed those comments. I changed where someone used "funct" instead of "func". Most notably I changed where the "offset" and "effAdd" variables were declared. They  are used in the SW and LW operations and were declared and used in each. I got errors for them being declared twice, and when declared in just one it gave a "jump to case label" error. So I moved the declaration to the beginning of the switch case. Now builds.